requirements.txt
ncclient==0.6.7
ruamel.yaml==0.16.6
pyang==2.1.1
lxml==4.4.2
netmiko==3.0.0
xmltodict
bpython
flake8
mypy
jinja2
pdbpp